refactor(actor api): extract iterate func
